Firmware: Add define for kernel DMA coherence.
If the kernel is configured with CONFIG_DMA_NONCOHERENT, then the DMA
API will return uncached(kseg1) buffers, or cached(kseg0) buffers with
CONFIG_DMA_COHERENT. The firmware must ensure to mirror this such that
the same (un)cached accesses are made, to ensure both sides see the up
to date data.
In reality, since this firmware is running on a CPU which must be
coherent with the rest of the system, once per-device DMA is supported
in the MIPS kernel then we can set the mips remote processor device to
coherent and assume that DMA_COHERENCE in the firmware is always 1.

Signed-off-by: Matt Redfearn <matt.redfearn@imgtec.com>